Moved routes into the src/ directory
