+++ package.json
@@ -44,6 +44,7 @@
     "graceful-fs": "^4.2.4",
     "html-loader": "^1.3.1",
     "mocha": "^6.2.3",
+    "nodemon": "^2.0.4",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
@@ -56,7 +57,8 @@
   },
   "scripts": {
     "build": "webpack --config configs/webpack.config.js --mode production",
-    "start": "node app.js",
+    "build:dev": "webpack --config configs/webpack.config.js --mode development",
+    "start": "npm run build:dev && npm test && node app.js",
     "test": "mocha --reporter spec --recursive"
   }
 }
