+++ test/routes/label-test.js
@@ -2,7 +2,7 @@ const { expect } = require('chai');
 const express = require('express');
 const request = require('supertest');
 
-const routes = require('./../../routes');
+const routes = require('./../../src/routes');
 
 describe('routes', () => {
   describe('label', () => {
